fix: Move link picker and metadata extraction to viewer to act in the user session aware iframe
