+++ src/document.js
@@ -2,8 +2,6 @@ import { emit } from '@nextcloud/event-bus'
 import { generateOcsUrl, getRootUrl, imagePath } from '@nextcloud/router'
 import { getRequestToken } from '@nextcloud/auth'
 import { loadState } from '@nextcloud/initial-state'
-import { showError } from '@nextcloud/dialogs'
-import { getLinkWithPicker } from '@nextcloud/vue/dist/Components/NcRichText.js'
 import Config from './services/config.tsx'
 import { setGuestName, shouldAskForGuestName } from './helpers/guestName.js'
 import { getUIDefaults, generateCSSVarTokens, getCollaboraTheme } from './helpers/coolParameters.js'
@@ -368,6 +366,9 @@ const documentsMain = {
 					'Action_FollowUser',
 					'Host_VersionRestore',
 					'Action_RemoveView',
+					'Action_InsertLink',
+					'Action_Paste',
+					'Action_GetLinkPreview_Resp',
 				]
 				PostMessages.registerPostMessageHandler(({ parsed, data }) => {
 					console.debug('[document] Received post message ', parsed)
@@ -496,12 +497,6 @@ const documentsMain = {
 						break
 					case 'UI_Mention':
 						documentsMain.sendUserList(parsed.args.text)
-						break
-					case 'UI_PickLink':
-						documentsMain.openLinkPicker()
-						break
-					case 'Action_GetLinkPreview':
-						documentsMain.resolveLink(args.url)
 						break
 					default:
 						console.debug('[document] Unhandled post message', parsed)
@@ -657,58 +652,6 @@ const documentsMain = {
 			return { username: user.label, profile }
 		})
 		PostMessages.sendWOPIPostMessage('loolframe', 'Action_Mention', { list })
-	},
-
-	async openLinkPicker() {
-		try {
-			const link = await getLinkWithPicker(null, true)
-			try {
-				const url = new URL(link)
-				if (url.protocol === 'http:' || url.protocol === 'https:') {
-					PostMessages.sendWOPIPostMessage('loolframe', 'Action_InsertLink', { url: link })
-					return
-				}
-			} catch (e) {
-				console.debug('error when parsing the link picker result')
-			}
-			PostMessages.sendWOPIPostMessage('loolframe', 'Action_Paste', { Mimetype: 'text/plain', Data: link })
-		} catch (e) {
-			showError(t('richdocuments', 'Failed to get a link with the picker'))
-			console.error('Link picker promise rejected :', e)
-		}
-	},
-
-	async resolveLink(url) {
-		try {
-			const result = await axios.get(generateOcsUrl('references/resolve', 2), {
-				params: {
-					reference: url,
-				},
-			})
-			const resolvedLink = result.data.ocs.data.references[url]
-			const title = resolvedLink?.openGraphObject?.name
-			const thumbnailUrl = resolvedLink?.openGraphObject?.thumb
-			if (thumbnailUrl) {
-				try {
-					const imageResponse = await axios.get(thumbnailUrl, { responseType: 'blob' })
-					if (imageResponse?.status === 200 && imageResponse?.data) {
-						const reader = new FileReader()
-						reader.addEventListener('loadend', (e) => {
-							const b64Image = e.target.result
-							PostMessages.sendWOPIPostMessage('loolframe', 'Action_GetLinkPreview_Resp', { url, title, image: b64Image })
-						})
-						reader.readAsDataURL(imageResponse.data)
-					}
-				} catch (e) {
-					console.error('Error loading the reference image', e)
-				}
-			} else {
-				PostMessages.sendWOPIPostMessage('loolframe', 'Action_GetLinkPreview_Resp', { url, title, image: null })
-			}
-		} catch (e) {
-			showError(t('richdocuments', 'Failed to get the link preview'))
-			console.error('Error resolving a reference', e)
-		}
 	},
 
 	onStartup() {
